Pass uint16_t vectors to vertex shader.

+++ TTauri/GUI/BackingPipeline_vulkan.vert
@@ -6,7 +6,7 @@ layout(push_constant) uniform PushConstants {
     vec2 viewportScale;
 } pushConstants;
 
-layout(location = 0) in vec3 inPosition;
+layout(location = 0) in uvec3 inPosition;
 layout(location = 1) in vec3 inAtlasPosition;
 layout(location = 2) in float inAlpha;
